requirements.txt
flask
flask-cors
flask-jwt-extended
python-dotenv
pymongo
bcrypt
python-multipart